refactor: LiveData 테스트용 확장 함수 분리
